pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agenticrag-learning-assistant"
version = "0.1.0"
description = "Agentic RAG learning assistant for C programming, operating systems, and cybersecurity labs."
readme = "PROJECT_STRUCTURE.md"
requires-python = ">=3.11,<3.13"
dependencies = [
  "Flask>=3.1,<4",
  "pydantic>=2.12,<3",
  "python-dotenv>=1.2,<2",
  "langchain>=1.2,<2",
  "langchain-core>=1.2,<2",
  "langchain-openai>=1.1,<2",
  "langgraph>=1.0,<2",
  "langgraph-checkpoint>=4.0,<5",
  "langmem>=0.0.30,<0.1",
  "lightrag-hku>=1.4.9.11,<2",
]

[project.optional-dependencies]
tools = [
  "neo4j>=6,<7",
  "networkx>=3.6,<4",
  "pyvis>=0.3,<1",
  "pypdf>=6.6,<7",
  "pipmaster>=1.1,<2",
]

[tool.setuptools]
py-modules = ["webapp"]
include-package-data = true

[tool.setuptools.packages.find]
include = ["agenticRAG*", "models*", "webapp_core*"]
